change threshold_double to float

+++ src/io/tree.cpp
@@ -46,7 +46,7 @@ Tree::~Tree() {
 }
 
 int Tree::Split(int leaf, int feature, int real_feature, uint32_t threshold_bin,
-                double threshold_double, double left_value, double right_value,
+                float threshold_double, double left_value, double right_value,
                 int left_cnt, int right_cnt, float gain, MissingType missing_type, bool default_left) {
   Split(leaf, feature, real_feature, left_value, right_value, left_cnt, right_cnt, gain);
   int new_node_idx = num_leaves_ - 1;
